[SymForce-External] Fix warnings
Fix warnings

Turn on -Wall, -Wextra, and -Werror, and fix all of them

Marking LCM C++ bindings, lcm_coretypes.h, generated test data (because
it contains LCM C++ bindings) as system includes so we don't fail the
build on warnings in them (which do exist).

The build should both now be warning-free (except for third-party
code), and fail if new warnings appear.

Reviewers: hayk-skydio,bradley-solliday-skydio,nathan-skydio
Topic: sym-warnings
Closes
GitOrigin-RevId: 8c6a26f8846e87c8a1905cabc7f9363e2c3f4226

+++ symforce/examples/robot_2d_localization/run_localization.cc
@@ -63,8 +63,8 @@ void RunLocalization() {
   values.Set({'d', 1}, 1.4);
   const std::array<std::array<double, 3>, 3> angles = {
       {{55, 245, -35}, {95, 220, -20}, {125, 220, -20}}};
-  for (int i = 0; i < angles.size(); ++i) {
-    for (int j = 0; j < angles[0].size(); ++j) {
+  for (int i = 0; i < static_cast<int>(angles.size()); ++i) {
+    for (int j = 0; j < static_cast<int>(angles[0].size()); ++j) {
       values.Set({'a', i, j}, angles[i][j] * M_PI / 180);
     }
   }
